refactor: create functions sub package
