+++ pkg/interpreter/interpreter_test.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"testing"
 
+	"github.com/jmespath-community/go-jmespath/pkg/functions"
 	"github.com/jmespath-community/go-jmespath/pkg/parsing"
 	"github.com/stretchr/testify/assert"
 )
@@ -50,7 +51,7 @@ func search(t *testing.T, expression string, data interface{}) (interface{}, err
 	if err != nil {
 		return nil, err
 	}
-	intr := NewInterpreter(nil)
+	intr := NewInterpreter(nil, functions.Default...)
 	return intr.Execute(ast, data)
 }
 
@@ -195,7 +196,7 @@ func TestCanSupportSliceOfStructsWithFunctions(t *testing.T) {
 
 func BenchmarkInterpretSingleFieldStruct(b *testing.B) {
 	assert := assert.New(b)
-	intr := NewInterpreter(nil)
+	intr := NewInterpreter(nil, functions.Default...)
 	parser := parsing.NewParser()
 	ast, _ := parser.Parse("fooasdfasdfasdfasdf")
 	data := benchmarkStruct{"foobarbazqux"}
@@ -209,7 +210,7 @@ func BenchmarkInterpretSingleFieldStruct(b *testing.B) {
 
 func BenchmarkInterpretNestedStruct(b *testing.B) {
 	assert := assert.New(b)
-	intr := NewInterpreter(nil)
+	intr := NewInterpreter(nil, functions.Default...)
 	parser := parsing.NewParser()
 	ast, _ := parser.Parse("fooasdfasdfasdfasdf.fooasdfasdfasdfasdf.fooasdfasdfasdfasdf.fooasdfasdfasdfasdf")
 	data := benchmarkNested{
@@ -234,7 +235,7 @@ func BenchmarkInterpretNestedMaps(b *testing.B) {
 	err := json.Unmarshal(jsonData, &data)
 	assert.Nil(err)
 
-	intr := NewInterpreter(nil)
+	intr := NewInterpreter(nil, functions.Default...)
 	parser := parsing.NewParser()
 	ast, _ := parser.Parse("fooasdfasdfasdfasdf.fooasdfasdfasdfasdf.fooasdfasdfasdfasdf.fooasdfasdfasdfasdf")
 	for i := 0; i < b.N; i++ {
